server/requirements.txt
# Server stack
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
pydantic>=2.6.0
pydantic-settings>=2.2.0

# Deep learning (install torch+cuda manually per your GPU)
# pip install torch torchvision --index-url https://download.pytorch.org/whl/cu121
torch>=2.0.0
torchvision>=0.15.0

# Vision models
ultralytics>=8.3.0
transformers>=4.40.0
timm>=0.9.0

# Image/video
opencv-python>=4.8.0
pillow>=10.0.0
ffmpeg-python>=0.2.0
supervision>=0.25.0
piexif>=1.1.3

# Numerical
numpy>=1.24.0
scipy>=1.11.0

# Utils
pyyaml>=6.0
tqdm>=4.66.0
python-dotenv>=1.0.0
